Set up Emotion provider and theme switcher toolbar button
