+++ .storybook/preview.tsx
@@ -1,0 +1,49 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import React from 'react';
+import type { Preview } from '@storybook/react';
+
+import { EuiProvider } from '../src/components/provider';
+
+const preview: Preview = {
+  decorators: [
+    (Story, context) => (
+      <EuiProvider colorMode={context.globals.colorMode}>
+        <Story />
+      </EuiProvider>
+    ),
+  ],
+  globalTypes: {
+    colorMode: {
+      description: 'Color mode for EuiProvider theme',
+      defaultValue: 'light',
+      toolbar: {
+        title: 'Color mode',
+        icon: 'circlehollow',
+        items: [
+          { value: 'light', title: 'Light mode', icon: 'circlehollow' },
+          { value: 'dark', title: 'Dark mode', icon: 'circle' },
+        ],
+        dynamicTitle: true,
+      },
+    },
+  },
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    backgrounds: { disable: true }, // Use colorMode instead
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+};
+
+export default preview;
